+++ src/biodwh2-core/src/main/java/de/unibi/agbi/biodwh2/core/etl/GraphExporter.java
@@ -9,6 +9,7 @@ import de.unibi.agbi.biodwh2.core.io.graph.GraphMLGraphWriter;
 import de.unibi.agbi.biodwh2.core.model.DataSourceFileType;
 import de.unibi.agbi.biodwh2.core.model.graph.*;
 import de.unibi.agbi.biodwh2.core.model.graph.meta.MetaGraph;
+import de.unibi.agbi.biodwh2.core.text.MetaGraphDynamicVisWriter;
 import de.unibi.agbi.biodwh2.core.text.MetaGraphStatisticsWriter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -58,11 +59,13 @@ public abstract class GraphExporter<D extends DataSource> {
     private void generateMetaGraphStatistics(final Workspace workspace, final Graph g) {
         final Path metaGraphImageFilePath = dataSource.getFilePath(workspace, DataSourceFileType.META_GRAPH_IMAGE);
         final Path metaGraphStatsFilePath = dataSource.getFilePath(workspace, DataSourceFileType.META_GRAPH_STATISTICS);
+        final Path metaGraphDynamicVisFilePath = dataSource.getFilePath(workspace, DataSourceFileType.META_GRAPH_DYNAMIC_VIS);
         if (workspace.getConfiguration().shouldSkipMetaGraphGeneration()) {
             if (LOGGER.isInfoEnabled())
                 LOGGER.info("Skipping '" + dataSource.getId() + "' meta graph generation as per configuration");
             FileUtils.safeDelete(metaGraphImageFilePath);
             FileUtils.safeDelete(metaGraphStatsFilePath);
+            FileUtils.safeDelete(metaGraphDynamicVisFilePath);
             return;
         }
         if (LOGGER.isInfoEnabled())
@@ -81,6 +84,8 @@ public abstract class GraphExporter<D extends DataSource> {
         final MetaGraphImage image = new MetaGraphImage(metaGraph, 1024, 1024);
         image.drawAndSaveImage(metaGraphImageFilePath);
         FileUtils.writeTextToUTF8File(metaGraphStatsFilePath, statistics);
+        final MetaGraphDynamicVisWriter visWriter = new MetaGraphDynamicVisWriter(metaGraph);
+        visWriter.write(metaGraphDynamicVisFilePath);
     }
 
     protected final <T> void createNodesFromModels(final Graph g, final Iterable<T> models) {